images: Get rid of the base image
Just build simx and kvm images. Now that the container version is gone
there is no reason for the extra layer.

This also means we have to rework how --pull is supported or it breaks,
make a version of --pull that does not depend on docker so doesn't have
the bug where it can't tell if something is a remote or local repository.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>