mkt: Provide man pages for the commands
This is done in a style similar to git where 'mkt cmd --help' will display
the man page instead of auto generated text.

Man pages are written in markdown and currently the client is required
to have pandoc installed to view the formatted version. Otherwise
the raw markdown test is shown.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>