images: Rework how the 'support' images are built
Instead of describing simx with a docker file that does git fetch/etc
handle the git fetch externally in the mkt script and inject a source tar
directly into docker. This allows us to fully cache the git fetches which
are very slow over VPN.

This is preparation for adding a bunch more built-from-source support
modules.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>