run: Stop qemu from clearing the screen and changing terminal settings
This is new behavior in 2.11, and it is done in the BIOS of all places.
Clearing the screen is annoying, but disabling line wrapping causes
kernel messages to be truncated and is not acceptable. Turn it off by
pointing the BIOS at a non existing serial port. It doesn't print anything
anyhow.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>