drvt: Add commands to launch drivertest outside the VM
Driver test has a GUI and a command line interface that needs to be run on
a host outside the VM to control the players.

Provide some simple wrappers that invoke the drivertest host programs
inside a container and plumb through X11 to the container so the GUI can
run.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>