run: Improve how the MAC address is determined
Speed up parsing the huge DHCP list file (by about 3x) and record the IP
and hostname from it, process the list in sorted order.

Pass this through to docker and do-kvm so that the VM's hostname is
correct.

Use docker to lock off the MACinstead of using the incredibly time
consuming 'ping'. MAC DAD in the booting kernel will inform the user
if they make a mistake

This makes 'mkt run' load the kernel basically instantly.

Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>