simx: Enable all supported by SimX Connect-4/5/6 devices
Honor users requests to enable specific device in SimX in any
combination, e.g. IB or ETH.

The SimX configuration file is generated on the fly in purpose.

It simplifies addition of extra options for the developers without
their need to rebuild images and for us removes the need to search
and replace previous values.

Signed-off-by: Leon Romanovsky <leonro@mellanox.com>