run: Execute custom boot script
Add new command line option --boot-script which receives path
to custom script. That script will be executed prior to providing
serial console to the user.

The same option is named "boot_script" and can be set in image sections
of configuration file.

Script is needed to be executable (+x bit) and is expected to have
shebang (#!...) line at the beginning.

Signed-off-by: Leon Romanovsky <leonro@mellanox.com>