run: Connect through SSH to already running container
The home users of MKT doesn't have br0 and runs its locally.
For them, the SSH client won't be able to connect to the QEMU image,
because the server is opened on localhost.

Change the code to ensure that "mkt run" executed after other container
is running will trigger SSH connection to that machine.

Signed-off-by: Leon Romanovsky <leonro@nvidia.com>